Add option to specify Core transformations to `internal core-eval`
