+++ app/Commands/Dev/Internal/CoreEval.hs
@@ -5,9 +5,11 @@ import Commands.Dev.Internal.CoreEval.Options
 import Data.HashMap.Strict qualified as HashMap
 import Evaluator
 import Juvix.Compiler.Core.Data.InfoTable
+import Juvix.Compiler.Core.Transformation qualified as Core
 import Juvix.Compiler.Core.Translation
 
 runCommand :: Members '[Embed IO, App] r => InternalCoreEvalOptions -> Sem r ()
 runCommand localOpts = do
   tab <- (^. coreResultTable) <$> runPipeline (localOpts ^. internalCoreEvalInputFile) upToCore
-  forM_ ((tab ^. infoMain) >>= ((tab ^. identContext) HashMap.!?)) (evalAndPrint localOpts tab)
+  let tab' = Core.applyTransformations (project localOpts ^. internalCoreEvalTransformations) tab
+  forM_ ((tab' ^. infoMain) >>= ((tab' ^. identContext) HashMap.!?)) (evalAndPrint localOpts tab')
